Rename DynamoDBTransactionLogSnapshotSaver
